% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkflagr-package.R
\docType{package}
\name{pkflagr-package}
\alias{pkflagr}
\alias{pkflagr-package}
\title{pkflagr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
