% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_find.R
\name{cm_flag}
\alias{cm_flag}
\title{Flag use of concomitant medication.}
\usage{
cm_flag(cm, drug_list = NULL)
}
\arguments{
\item{cm}{Concomitant medication table as data frame.}

\item{drug_list}{A data frame with the drugs to be identified, with the columns
DRUG, TYPE, QUALIFIER and TARGET.}
}
\description{
Flag use of concomitant medication.
}
