% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_find.R
\name{make_ddi_drugs}
\alias{make_ddi_drugs}
\title{Make list of DDI perpetrators or substrates}
\usage{
make_ddi_drugs(
  drug_list = fdi_clinical_ddi_drug_list,
  type = c("inhibitor", "substrate", "inducer"),
  target = c("2D6", "1A2", "OAT1", "3A4", "2C9", "P-gp", "2C19", "OATP1B1", "OATP1B3",
    "OAT3", "2B6", "MATE1", "MATE2-K", "2C8", "BCRP")
)
}
\arguments{
\item{drug_list}{The list of drugs, defaults to a built-in data set provided
by FDA. Must contain columns DRUG, TYPE, and TARGET.}

\item{type}{Type of DDI perpetrator or substrate, as character vector.
Valid values are "inhibitor", "substrate", and "inducer".}

\item{target}{Type of DDI target, as character vector. Valid values are
"2D6", "1A2", "OAT1", "3A4", "2C9", "P-gp", "2C19", "OATP1B1",
"OATP1B3", "OAT3", "2B6", "MATE1", "MATE2-K", "2C8", "BCRP".}
}
\value{
A data frame containing the filtered drug list with columns:
\itemize{
\item DRUG: The canonical drug name (uppercase, without numbers and commas)
\item TYPE: The type of DDI (inhibitor, substrate, or inducer)
\item TARGET: The target of the DDI
\item QUALIFIER: The intensity qualifier (if present in input)
}
}
\description{
Make list of DDI perpetrators or substrates
}
\examples{
make_ddi_drugs(type = "inhibitor", target = "2B6")
}
